+++ tests/AuthorizationInterceptor.Tests/Utils/MockAuthorizationInterceptor.cs
@@ -5,12 +5,12 @@ namespace AuthorizationInterceptor.Tests.Utils;
 
 public class MockAuthorizationInterceptor : IAuthorizationInterceptor
 {
-    public ValueTask<AuthorizationHeaders?> GetHeadersAsync(string name, CancellationToken cancellationToken, string? cacheKeySuffix = null)
+    public ValueTask<AuthorizationHeaders?> GetHeadersAsync(string key, CancellationToken cancellationToken)
     {
         return ValueTask.FromResult<AuthorizationHeaders?>(null);
     }
 
-    public ValueTask UpdateHeadersAsync(string name, AuthorizationHeaders? expiredHeaders, AuthorizationHeaders? newHeaders, CancellationToken cancellationToken, string? cacheKeySuffix = null)
+    public ValueTask UpdateHeadersAsync(string key, AuthorizationHeaders? expiredHeaders, AuthorizationHeaders? newHeaders, CancellationToken cancellationToken)
     {
         return ValueTask.CompletedTask;
     }